mads testing streamlined

+++ examples/bayesian_sampling/runtests.jl
@@ -1,6 +1,6 @@
 import Mads
 
-info("Bayesian sampling ...")
+Mads.madsinfo("Bayesian sampling ...")
 
 workdir = Mads.getmadsdir() # get the directory where the problem is executed
 if workdir == ""
@@ -11,26 +11,20 @@ md = Mads.loadmadsfile(workdir * "internal-linearmodel.mads")
 rootname = Mads.getmadsrootname(md)
 mcmcchain = Mads.bayessampling(md; nsteps=10, burnin=1, thinning=1, seed=2016)
 Mads.savemcmcresults(mcmcchain.value', rootname * "-test-mcmcchain1.json")
-if !haskey(ENV, "MADS_NO_PLOT")
-	#Mads.scatterplotsamples(md, mcmcchain.value', rootname * "-test-mcmcchain1.svg")
-end
 
 mcmcchains = Mads.bayessampling(md, 2; nsteps=10, burnin=1, thinning=1, seed=2016)
 values = vcat(map(chain->chain.value', mcmcchains)...)
-if !haskey(ENV, "MADS_NO_PLOT")
-	#Mads.scatterplotsamples(md, values, rootname * "-test-mcmcchain3.svg")
-end
 
 md = Mads.loadmadsfile(workdir * "w01.mads")
 rootname = Mads.getmadsrootname(md)
 mcmcchain = Mads.bayessampling(md; nsteps=10, burnin=1, thinning=1, seed=2016)
-if !haskey(ENV, "MADS_NO_PLOT")
-	#Mads.scatterplotsamples(md, mcmcchain.value', rootname * "-test-bayes-results.svg")
+if !haskey(ENV, "MADS_NO_PLOT") && Mads.long_tests
+	Mads.scatterplotsamples(md, mcmcchain.value', rootname * "-test-bayes-results.svg")
 	mcmcvalues = Mads.paramarray2dict(md, mcmcchain.value') # convert the parameters in the chain to a parameter dictionary of arrays
-	#Mads.spaghettiplots(md, mcmcvalues, keyword="test")
-	#Mads.spaghettiplot(md, mcmcvalues, keyword="test")
-	#Mads.spaghettiplots(md, 3, keyword="test")
-	#Mads.spaghettiplot(md, 3, keyword="test")
+	Mads.spaghettiplots(md, mcmcvalues, keyword="test")
+	Mads.spaghettiplot(md, mcmcvalues, keyword="test")
+	Mads.spaghettiplots(md, 3, keyword="test")
+	Mads.spaghettiplot(md, 3, keyword="test")
 	s = rootname
 	run(`bash -c "rm -f $(s)*-test-*.svg"`)
 end
